refactor : 피드백 받은 사항 변경 (#75)
* refactor : 피드백 받은 사항 변경
- 구매 도메인에 Swagger 추가
- OrderService.java에 주석 추가
- OrderService.java에 중복 save 제거
- 구매 후 포인트 깎이는 것 적용 (User에 substractPoint()함수 추가)
- 최신 develop 적용 이후 H2 웹페이지가 잘 나오도록 JwtAuthenticationFilter.java에 H2 예외 추가 (피드백 요망)
- @CookieValue("userId")를 암호화하는 것은 좀 더 조사해보겠습니다.

* refactor : 피드백 받은 사항 변경
- JwtAuthenticationFilter.java 원상복귀

+++ backend/src/main/java/com/backend/petplace/domain/order/entity/Order.java
@@ -3,8 +3,6 @@ package com.backend.petplace.domain.order.entity;
 import com.backend.petplace.domain.orderproduct.entity.OrderProduct;
 import com.backend.petplace.domain.user.entity.User;
 import com.backend.petplace.global.entity.BaseEntity;
-import com.backend.petplace.global.exception.BusinessException;
-import com.backend.petplace.global.response.ErrorCode;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
@@ -41,7 +39,7 @@ public class Order extends BaseEntity {
 
   @NotNull
   @Min(0)
-  private Long totalPrice;
+  private Integer totalPrice;
 
   @NotNull
   @Enumerated(EnumType.STRING)
@@ -51,14 +49,14 @@ public class Order extends BaseEntity {
   private List<OrderProduct> orderProducts = new ArrayList<>();
 
   @Builder
-  public Order(User user, Long totalPrice, OrderStatus orderStatus) {
+  public Order(User user, Integer totalPrice, OrderStatus orderStatus) {
     this.user = user;
     this.totalPrice = totalPrice;
     this.orderStatus = orderStatus;
   }
 
   //정적 팩토리 메서드를 통한 Order 객체 생성
-  public static Order createOrder(User user, Long totalPrice) {
+  public static Order createOrder(User user, Integer totalPrice) {
     return Order.builder()
         .user(user)
         .totalPrice(totalPrice)
